+++ src/main/java/com/github/mauricioaniche/ck/metric/composed/ProgramLevel.java
@@ -20,7 +20,7 @@ public class ProgramLevel implements ComposedMetric {
 		for (CKNumber ck : report.all()) {
 			double pLevel = 0.0;
 			try {
-				pLevel = (double) 1 / ck.getDifficultLevel();
+				pLevel = 1 / ck.getDifficultLevel();
 			} catch (Exception e) {
 				pLevel = 0.0;
 			}
